:recycle: Refactor `load_torch_model` and `model_to` to models_abc.py
